+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/indispensable/pki/DistinguishedNameTest.kt
@@ -1,6 +1,6 @@
 package at.asitplus.signum.indispensable.pki
 
-import at.asitplus.signum.indispensable.asn1.KnownOIDs
+import at.asitplus.signum.indispensable.asn1.*
 import at.asitplus.test.FreeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe